Rest controllers

+++ src/main/java/com/revature/koality/service/PostReviewServiceImpl.java
@@ -7,14 +7,18 @@ import org.springframework.stereotype.Service;
 import com.revature.koality.bean.AlbumReview;
 import com.revature.koality.bean.ReviewContent;
 import com.revature.koality.bean.TrackReview;
+import com.revature.koality.dao.AlbumDAO;
 import com.revature.koality.dao.AlbumReviewDAO;
 import com.revature.koality.dao.AlbumReviewDAOImpl;
+import com.revature.koality.dao.TrackDAO;
 import com.revature.koality.dao.TrackReviewDAO;
 import com.revature.koality.dao.TrackReviewDAOImpl;
 
 @Service("postReviewServiceImpl")
 public class PostReviewServiceImpl implements PostReviewService {
 
+	private TrackDAO td;
+	private AlbumDAO ad;
 	private TrackReviewDAO trd;
 	private AlbumReviewDAO ard;
 
@@ -30,6 +34,26 @@ public class PostReviewServiceImpl implements PostReviewService {
 	public PostReviewServiceImpl(AlbumReviewDAOImpl ard) {
 		super();
 		this.ard = ard;
+	}
+
+	public TrackDAO getTd() {
+		return td;
+	}
+
+	@Autowired
+	@Qualifier("trackDAOImpl")
+	public void setTd(TrackDAO td) {
+		this.td = td;
+	}
+
+	public AlbumDAO getAd() {
+		return ad;
+	}
+
+	@Autowired
+	@Qualifier("albumDAOImpl")
+	public void setAd(AlbumDAO ad) {
+		this.ad = ad;
 	}
 
 	public TrackReviewDAO getTrd() {
@@ -55,39 +79,67 @@ public class PostReviewServiceImpl implements PostReviewService {
 	@Override
 	public int postTrackReview(int rating, String reviewComment, int trackId, int customerId) {
 
-		ReviewContent reviewContent = new ReviewContent(rating, reviewComment);
-
-		return trd.addTrackReview(reviewContent, trackId, customerId);
+		if (!td.hasAccessAsCustomer(trackId, customerId)) {
+			return 0;
+		} else {
+			ReviewContent reviewContent = new ReviewContent(rating, reviewComment);
+			return trd.addTrackReview(reviewContent, trackId, customerId);
+		}
 
 	}
 
 	@Override
 	public int postAlbumReview(int rating, String reviewComment, int albumId, int customerId) {
 
-		ReviewContent reviewContent = new ReviewContent(rating, reviewComment);
-
-		return ard.addAlbumReview(reviewContent, albumId, customerId);
+		if (!ad.hasAccessAsCustomer(albumId, customerId)) {
+			return 0;
+		} else {
+			ReviewContent reviewContent = new ReviewContent(rating, reviewComment);
+			return ard.addAlbumReview(reviewContent, albumId, customerId);
+		}
 
 	}
 
 	@Override
 	public TrackReview viewPostedTrackReview(int trackId, int customerId) {
-		return trd.getTrackReviewByTrackIdAndCustomerId(trackId, customerId);
+
+		TrackReview trackReview = trd.getTrackReviewByTrackIdAndCustomerId(trackId, customerId);
+		trackReview.setTrack(null);
+		trackReview.setCustomer(null);
+
+		return trackReview;
 	}
 
 	@Override
 	public AlbumReview viewPostedAlbumReview(int albumId, int customerId) {
-		return ard.getAlbumReviewByAlbumIdAndCustomerId(albumId, customerId);
+
+		AlbumReview albumReview = ard.getAlbumReviewByAlbumIdAndCustomerId(albumId, customerId);
+		albumReview.setAlbum(null);
+		albumReview.setCustomer(null);
+
+		return albumReview;
 	}
 
 	@Override
-	public boolean deleteTrackReview(int trackReviewId) {
-		return trd.deleteTrackReview(trackReviewId);
+	public boolean deleteTrackReview(int trackReviewId, int customerId) {
+
+		if (trd.isOwner(trackReviewId, customerId)) {
+			return trd.deleteTrackReview(trackReviewId);
+		} else {
+			return false;
+		}
+
 	}
 
 	@Override
-	public boolean deleteAlbumReview(int albumReviewId) {
-		return ard.deleteAlbumReview(albumReviewId);
+	public boolean deleteAlbumReview(int albumReviewId, int customerId) {
+
+		if (ard.isOwner(albumReviewId, customerId)) {
+			return ard.deleteAlbumReview(albumReviewId);
+		} else {
+			return false;
+		}
+
 	}
 
 }
